requirements.txt
absl-py==2.1.0
asgiref==3.7.2
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
djangorestframework==3.15.1
dm-tree==0.1.8
flatbuffers==24.3.7
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
keras==3.0.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pillow==10.2.0
protobuf==4.25.3
Pygments==2.17.2
requests==2.31.0
rich==13.7.1
six==1.16.0
sqlparse==0.4.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
